:green_heart: renamed server.js
renamed server.js because I think it's breaking the azure deployment
