+++ packages/server/src/index.ts
@@ -1,4 +1,4 @@
-import app from "./server";
+import app from "./express-server";
 import dotenv from "dotenv";
 
 dotenv.config();
